backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
numpy
scikit-learn
scipy
python-dotenv
requests
httpx